requirement.txt
torch==2.0.0
torchvision==0.15.1
matplotlib==3.7.1
wandb==0.15.4
hydra-core==1.3.2
scipy==1.10.1
matplotlib-inline==0.1.6
ipdb==0.13.13
tqdm==4.65.0